+++ doc/release/release_notes.rst
@@ -43,6 +43,12 @@ Release Notes
         :class:`ftrack_api.data.String` is now using a temporary file instead of
         StringIO to avoid reading large files into memory.
 
+    .. change:: fixed
+        :tags: session, locations
+
+        `ftrack.centralized-storage` does not properly validate location
+        selection during user configuration.
+
 .. release:: 0.16.0
     :date: 2016-10-18